Moving test_issue_* tests to more appropriate locations. All SchemaView tests should now be in the test_schemaview.py file.
