+++ tests/test_issues/test_include_schema.py
@@ -4,6 +4,8 @@ from linkml_runtime.linkml_model.meta import SchemaDefinition
 from linkml_runtime.loaders import yaml_loader
 from tests.test_issues.environment import env
 
+# https://github.com/linkml/linkml/issues/3
+
 
 class IncludeSchemaTestCase(unittest.TestCase):
     """include_schema.yaml produces a Python exception on an uncaught error"""
