+++ tests/test_issues/test_issue_368_enums.py
@@ -4,6 +4,8 @@ from linkml_runtime.dumpers import json_dumper, rdf_dumper, yaml_dumper
 from linkml_runtime.utils.compile_python import compile_python
 from tests.test_issues.environment import env
 
+# https://github.com/linkml/linkml/issues/368
+
 
 def header(txt: str) -> str:
     return "\n" + ("=" * 20) + f" {txt} " + ("=" * 20)
