+++ tests/test_issues/test_issue_8.py
@@ -5,6 +5,8 @@ from linkml_runtime.loaders import yaml_loader
 from linkml_runtime.utils.yamlutils import YAMLRoot
 from tests.test_issues.environment import env
 
+# https://github.com/linkml/linkml/issues/8
+
 
 def override(cls: type[YAMLRoot]):
     orig = cls.MissingRequiredField
