+++ tests/test_loaders_dumpers/input/issue_1040.yaml
