+++ tests/test_loaders_dumpers/input/issue_576.yaml
