+++ tests/test_loaders_dumpers/test_rdflib_dumper.py
@@ -1,27 +1,26 @@
+import logging
 import os
 import unittest
-import logging
 from pathlib import Path
 
 from curies import Converter
-from rdflib import Graph, Literal, URIRef
+from rdflib import Graph, Literal, Namespace, URIRef
 from rdflib.namespace import RDF, SKOS, XSD
-from rdflib import Namespace
 
-from linkml_runtime import MappingError, DataNotFoundError
+from linkml_runtime import DataNotFoundError, MappingError
 from linkml_runtime.dumpers import rdflib_dumper, yaml_dumper
 from linkml_runtime.linkml_model import Prefix
-from linkml_runtime.loaders import yaml_loader
-from linkml_runtime.loaders import rdflib_loader
+from linkml_runtime.loaders import rdflib_loader, yaml_loader
 from linkml_runtime.utils.schemaview import SchemaView
 from tests.test_loaders_dumpers import INPUT_DIR, OUTPUT_DIR
-from tests.test_loaders_dumpers.models.personinfo import Container, Person, Address, Organization, OrganizationType
 from tests.test_loaders_dumpers.models.node_object import NodeObject, Triple
+from tests.test_loaders_dumpers.models.personinfo import Address, Container, Organization, OrganizationType, Person
+from tests.test_loaders_dumpers.models.personinfo_test_issue_429 import Container as Container_429
 from tests.test_loaders_dumpers.models.phenopackets import (
-    PhenotypicFeature,
+    MetaData,
     OntologyClass,
     Phenopacket,
-    MetaData,
+    PhenotypicFeature,
     Resource,
 )
 
@@ -82,15 +81,15 @@ P:001 a sdo:Person ;
     sdo:email "fred.bloggs@example.com" ;
     sdo:name "fred bloggs" ;
     personinfo:age_in_years 33 ;
-    personinfo:has_interpersonal_relationships [ 
+    personinfo:has_interpersonal_relationships [
       a personinfo:InterpersonalRelationship ;
       personinfo:type famrel:70 ;
-      personinfo:related_to P:002 
+      personinfo:related_to P:002
     ] ,
-    [ 
+    [
       a personinfo:InterpersonalRelationship ;
       personinfo:type "BEST_FRIEND_OF" ;
-      personinfo:related_to P:003 
+      personinfo:related_to P:003
     ] .
 """
 
@@ -442,6 +441,46 @@ class RDFLibConverterDumperTestCase(RdfLibDumperTestCase):
     def setUp(self) -> None:
         """Set up the test case using a :class:`curies.Converter` instead of a simple prefix map."""
         self.prefix_map = Converter.from_prefix_map(PREFIX_MAP)
+
+
+class RdfLibPrefixTestCase(unittest.TestCase):
+    # see https://github.com/linkml/linkml/issues/429
+
+    SCHEMA = os.path.join(INPUT_DIR, "personinfo_test_issue_429.yaml")
+    DATA = os.path.join(INPUT_DIR, "example_personinfo_test_issue_429_data.yaml")
+    OUT = os.path.join(OUTPUT_DIR, "example_personinfo_test_issue_429_data.ttl")
+
+    ORCID = Namespace("https://orcid.org/")
+    personinfo = Namespace("https://w3id.org/linkml/examples/personinfo/")
+    SDO = Namespace("http://schema.org/")
+
+    def setUp(self):
+        self.g = self.create_rdf_output()
+
+    def create_rdf_output(self):
+        view = SchemaView(self.SCHEMA)
+        container = yaml_loader.load(self.DATA, target_class=Container_429)
+        rdflib_dumper.dump(container, schemaview=view, to_file=self.OUT)
+        g = Graph()
+        g.parse(self.OUT, format="ttl")
+        return g
+
+    def test_rdf_output(self):
+        self.assertIn((self.ORCID["1234"], RDF.type, self.SDO.Person), self.g)
+        self.assertIn((self.ORCID["1234"], self.personinfo.full_name, Literal("Clark Kent")), self.g)
+        self.assertIn((self.ORCID["1234"], self.personinfo.age, Literal("32")), self.g)
+        self.assertIn((self.ORCID["1234"], self.personinfo.phone, Literal("555-555-5555")), self.g)
+        self.assertIn((self.ORCID["4567"], RDF.type, self.SDO.Person), self.g)
+        self.assertIn((self.ORCID["4567"], self.personinfo.full_name, Literal("Lois Lane")), self.g)
+        self.assertIn((self.ORCID["4567"], self.personinfo.age, Literal("33")), self.g)
+        self.assertIn((self.ORCID["4567"], self.personinfo.phone, Literal("555-555-5555")), self.g)
+
+    def test_output_prefixes(self):
+        with open(self.OUT, encoding="UTF-8") as file:
+            file_string = file.read()
+        prefixes = ["prefix ORCID:", "prefix personinfo:", "prefix sdo:", "sdo:Person", "personinfo:age", "ORCID:1234"]
+        for prefix in prefixes:
+            assert prefix in file_string
 
 
 if __name__ == "__main__":
