+++ tests/test_utils/input/issue_1126.yaml
