+++ tests/test_utils/input/issue_590.yaml
@@ -1,5 +1,5 @@
 id: https://w3id.org/
-name: linkml_issue_590
+name: issue_590
 
 classes:
   foo:
