+++ tests/test_utils/test_metamodelcore.py
@@ -1,7 +1,7 @@
 import datetime
 import unittest
 from dataclasses import dataclass
-
+import pytest
 from jsonasobj2 import as_json
 from rdflib import Literal, XSD, Graph, RDF, Namespace
 
@@ -297,6 +297,20 @@ class MetamodelCoreTest(unittest.TestCase):
         )
         with self.assertRaises(ValueError):
             y = Pair(s, s)
+
+
+def test_issue_1355_invalid_url_message() -> None:
+    """Check that quotes are used when referencing invalid urls to improve troubleshooting UX.
+
+    See https://github.com/linkml/linkml/issues/1355, improve invalid URL message
+    """
+    #  note the trailing blank
+    url = "https://ceur-ws.org/Vol-2931/ICBO_2019_paper_20.pdf "
+    with pytest.raises(
+        ValueError,
+        match="'https://ceur-ws.org/Vol-2931/ICBO_2019_paper_20.pdf ': is not a valid URI",
+    ):
+        URI(url)
 
 
 if __name__ == "__main__":
